requirements.txt
pymongo[srv]==4.1.1
certifi
xgboost
catboost
pandas
numpy
seaborn
matplotlib
scikit-learn
imblearn
miceforest
kneed
prettytable
python-dotenv
PyYAML
-e .